Win32API 강의 42화. Win32API Camera (3)

+++ Client/CAnimation.cpp
@@ -75,11 +75,11 @@ void CAnimation::Create(CTexture* _pTex, Vec2 _vLT, Vec2 _vSliceSize, Vec2 _vSte
 	m_pTex = _pTex;
 
 	tAnimFrm frm = {};
-	for (int i = 0; i < _iFreamCount; ++i)
+	for (UINT i = 0; i < _iFreamCount; ++i)
 	{
 		frm.fDuration = _fDuration;
 		frm.vSlice = _vSliceSize;
-		frm.vLT = _vLT + _vStep * i;
+		frm.vLT = _vLT + _vStep * (float)i;
 
 		m_vecFrm.push_back(frm);
 	}
